Removed unnecessary packages, configs and renamed src files
