+++ packages/arm-batch-rest/config/tsconfig.common.json
@@ -3,9 +3,7 @@
   "compilerOptions": {
     "outDir": "../lib",
     "rootDir": "../src/",
-    "typeRoots": ["../node_modules/@types"],
-    "noUnusedLocals": false,
-    "noUnusedParameters": false
+    "typeRoots": ["../node_modules/@types"]
   },
   "include": ["../src/"]
 }
