+++ packages/arm-batch-rest/package.json
@@ -81,29 +81,11 @@
   },
   "devDependencies": {
     "@batch/common-config": "^1.0.0",
-    "@azure/identity": "^2.0.1",
-    "@types/chai": "^4.2.8",
     "@types/jest": "^29.2.4",
     "@types/node": "^14.0.0",
-    "@types/node-fetch": "^2.6.2",
-    "autorest": "latest",
-    "chai": "^4.2.0",
+    "autorest": "3.6.3",
     "jest": "^29.3.1",
     "jest-junit": "^12.2.0",
-    "mkdirp": "^1.0.4",
-    "node-fetch": "^2.6.7",
-    "nyc": "^14.0.0",
-    "source-map-support": "^0.5.9",
-    "ts-jest": "^29.0.3",
-    "ts-node": "^10.9.1",
-    "typescript": "~4.9.4"
-  },
-  "//metadata": {
-    "constantPaths": [
-      {
-        "path": "swagger/README.md",
-        "prefix": "package-version"
-      }
-    ]
+    "ts-jest": "^29.0.3"
   }
 }
