+++ packages/arm-batch-rest/src/__tests__/setup-tests.ts
@@ -1,12 +1,5 @@
 import { destroyEnvironment } from "@batch/ui-common";
-import fetch, { Headers, Request, Response } from "node-fetch";
-
-if (!("fetch" in globalThis)) {
-    Object.assign(globalThis, { fetch, Headers, Request, Response });
-}
 
 afterEach(() => {
     destroyEnvironment();
 });
-
-jest.setTimeout(10000);
