+++ packages/arm-batch-rest/src/http/pipeline-http-headers.ts
