+++ packages/arm-batch-rest/src/http/rest-error.ts
